Deduplicate code for installation URL/path formatting
