+++ src/components/installer/installationlayout.ts
@@ -1,0 +1,25 @@
+import { Platform } from "../../shell";
+
+export function platformUrlString(platform: Platform, supported?: Platform[]): string | null {
+    if (supported && supported.indexOf(platform) < 0) {
+        return null;
+    }
+    switch (platform) {
+        case Platform.Windows: return 'windows';
+        case Platform.MacOS: return 'darwin';
+        case Platform.Linux: return 'linux';
+        default: return null;
+    }
+}
+
+export function formatBin(tool: string, platform: Platform): string | null {
+    const platformString = platformUrlString(platform);
+    if (!platformString) {
+        return null;
+    }
+    const toolPath = `${platformString}-amd64/${tool}`;
+    if (platform === Platform.Windows) {
+        return toolPath + '.exe';
+    }
+    return toolPath;
+}
